+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.7] - 2025-05-28
+
+### Added
+
+- Patterns for `:matches-property()` and generic CSS rules.
+
+### Changed
+
+- Exclude `$removeheader`, `:matches-property()` and generic CSS rules from uBO filter lists.
+
+### Fixed
+
+- Patterns for CSS rules.
+
+[v1.2.7]: https://github.com/AdguardTeam/FiltersCompiler/compare/v1.2.6...v1.2.7
+
 ## [v1.2.6] - 2025-04-29
 
 ### Added